+++ Echo/src/utopia/echo/controller/chat/BufferedReplyGenerator.scala
@@ -1,21 +1,23 @@
 package utopia.echo.controller.chat
 
+import utopia.annex.model.response.RequestResult
+
 import scala.concurrent.Future
 import scala.language.implicitConversions
-import scala.util.Try
 
 object BufferedReplyGenerator
 {
 	// IMPLICIT ----------------------
 	
-	implicit def apply[R](f: String => Future[Try[R]]): BufferedReplyGenerator[R] = new _BufferedReplyGenerator[R](f)
+	implicit def apply[R](f: String => Future[RequestResult[R]]): BufferedReplyGenerator[R] =
+		new _BufferedReplyGenerator[R](f)
 	
 	
 	// NESTED   ----------------------
 	
-	private class _BufferedReplyGenerator[+R](f: String => Future[Try[R]]) extends BufferedReplyGenerator[R]
+	private class _BufferedReplyGenerator[+R](f: String => Future[RequestResult[R]]) extends BufferedReplyGenerator[R]
 	{
-		override def bufferedReplyFor(message: String): Future[Try[R]] = f(message)
+		override def bufferedReplyFor(message: String): Future[RequestResult[R]] = f(message)
 	}
 }
 
@@ -33,5 +35,5 @@ trait BufferedReplyGenerator[+R]
 	 * @param message A message / prompt to send
 	 * @return A future that yields the reply, if successful
 	 */
-	def bufferedReplyFor(message: String): Future[Try[R]]
+	def bufferedReplyFor(message: String): Future[RequestResult[R]]
 }