+++ Echo/src/utopia/echo/model/enumeration/MessageStopReason.scala
@@ -6,6 +6,19 @@ package utopia.echo.model.enumeration
  * @since 29.01.2026, v1.5
  */
 sealed trait MessageStopReason
+{
+	// ABSTRACT -----------------------
+	
+	/**
+	 * @return Key used for representing this stop reason
+	 */
+	def key: String
+	
+	
+	// IMPLEMENTED  -------------------
+	
+	override def toString = key
+}
 
 object MessageStopReason
 {
@@ -23,20 +36,35 @@ object MessageStopReason
 	 * Indicates that a stop sequence was encountered, and the message completed naturally.
 	 */
 	case object MessageCompleted extends MessageStopReason
+	{
+		override val key: String = "stop"
+	}
 	/**
 	 * Indicates that the maximum context size was reached
 	 */
 	case object LengthLimitReached extends MessageStopReason
+	{
+		override val key: String = "length"
+	}
 	/**
 	 * Indicates that the message stopped in order to call a tool
 	 */
 	case object ToolCalled extends MessageStopReason
+	{
+		override val key: String = "tool_calls"
+	}
 	/**
 	 * Indicates that a message was blocked by server-side content moderation mechanism
 	 */
 	case object Censored extends MessageStopReason
+	{
+		override val key: String = "content_filter"
+	}
 	/**
 	 * Indicates that the server was unable to complete the message for technical reasons
 	 */
 	case object ServerError extends MessageStopReason
+	{
+		override val key: String = "error"
+	}
 }